+++ src/util/bw-png.c
@@ -43,8 +43,8 @@ static const char RCSid[] = "@(#)$Header$ (BRL)";
 #include "bn.h"
 #include "zlib.h"
 
-static int	file_width = 512;		/* default input width */
-static int	file_height = 512;		/* default input height */
+static long int	file_width = 512;		/* default input width */
+static long int	file_height = 512;		/* default input height */
 static int	autosize = 0;			/* !0 to autosize input */
 static int	fileinput = 0;			/* file of pipe on input? */
 static char	*file_name;
@@ -71,11 +71,11 @@ get_args(int argc, register char **argv)
 			break;
 		case 's':
 			/* square file size */
-			file_height = file_width = atoi(optarg);
+			file_height = file_width = atol(optarg);
 			autosize = 0;
 			break;
 		case 'w':
-			file_width = atoi(optarg);
+			file_width = atol(optarg);
 			autosize = 0;
 			break;
 		case 'n':
@@ -127,10 +127,10 @@ main(int argc, char **argv)
 
 	/* autosize input? */
 	if( fileinput && autosize ) {
-		int	w, h;
+		unsigned long int	w, h;
 		if( bn_common_file_size(&w, &h, file_name, 1) ) {
-			file_width = w;
-			file_height = h;
+			file_width = (long)w;
+			file_height = (long)h;
 		} else {
 			fprintf(stderr,"bw-png: unable to autosize\n");
 		}
